requirements.txt
scipy 
numpy 
matplotlib
pandas
scikit-learn
vllm
transformers
mlflow
PyGSP
tyro
networkx
accelerate
cattrs
datasets
wandb
deepspeed
trl
backoff
nltk
torch
PyYAML
git+https://github.com/EltayebAhmed/gllm.git
